moving config_parse.py to be in code base, changing init import to match
